[MapIntrinToIml] Don't treat XMM16-31 as callee-saved registers for 256-bit masked SVML functions. (#10077) (#10344)
SVML calls can use svml_avx_avx_impl_cc for more callee-saved registers
if the variant to be called is guarenteed to be an AVX one. But recently
we've found masked functions may reroute the call to its non-masked
counterpart with dynamic dispatch. Until it's fixed we can't apply this
optimization for masked functions.

JIRA: CMPLRLLVM-43721