[LoopOpt][SafeRedn] Use reassoc vs fast flag for conditionally updated FP reductions during marking of unsafe algebra (#10663)
* [LoopOpt][SafeRedn] Use reassoc vs fast flag for conditionally updated FP reductions during marking of unsafe algebra

Code in getConditionalAndUnsafeAlgebraInfo was using fast flag on FP instructions to avoid
marking reductions with HasUnsafeAlgebra flag. This was the only remaining place in loopopt
code that was using fast flag and blocked the effort to replace checks for fast flag with
checks for appropriate flags. In this case, a check for AllowReassoc flag is sufficient.
Earlier conditional path was using a check for fast flag and unconditional path used a check
for reassoc flag. This PR modifies both conditional and unconditional paths to use reassoc
flag while setting HasUnsafeAlgebra.

Jira :: CMPLRLLVM-22580

* Address Pankaj's comment about updating safe reduction lit test