[NFC] Make dummy changes to LIT tests to restore them in xmain-web. (#10750)
These tests were inadvertantly deleted as part of the conflict
resolution of #10568 in xmain-web ($10690).

If we add them back directly to xmain-web, we will lose their git
blame history. So we are making these dummy changes here, which
which will lead to a conflict in xmain-web (because xmain-web don't
have these files anymore). As part of that conflict's resolution,
we can keep the incoming files/changes from xmain, ignoring the
deleted files from xmain-web.

JIRA: CMPLRLLVM-44758