[OpenCL] CMPLRLLVM-44272 Add missing OpenCL extensions to SATest (#10621)
OpenCL extensions are defined in cpu_device module. The extensions
definition queries the device properties. It is probably better to stay in
cpu_device module instead of in utility files that could be shared between
cpu_device module and SATest.

Re-defining extensions within backend/validations so that there is no
circular dependency between cpu_device and backend. The extensions
doesn't change frequently, so we can update extensions on demand when
a SATest fails.