[VPlan] Fix up incomplete reduction in inner loop (#10851)
[VPlan] Fix up incomplete reduction in inner loop

A customer encountered a compfail in a case where a reduction description
was incomplete and needed to be completed by VPlan.  The code to do this
didn't anticipate the possibility of the reduction operation occurring at
a deeper level than the loop where the reduction was specified.  This
patch fixes the oversight.  Code fix by Karthik Senthil; reduced test case
is mine.

Reference: CMPLRLLVM-44206