[Paropt] Improve renaming of constexpr clause operands. (#10429)
Previously, when the clause operand was a constexpr, like:

```
  [4 x i8]* bitcast (i8* getelementptr inbounds ([20 x i8]* @global,
                                                 i32 0, i32 4)
                     to [4 x i8]*)
```

while the use of the inside the body of the function looked like:

   `i32* bitcast (i8* getelementptr inbounds ([20 x i8]* @global,
                                              i32 0, i32 4))
                  to i32*
```

Then we were unable to match the two constexprs since they are not
identical. Now we try to ignore any no-op casts like zero-offset GEPs
and bitcasts and do the replacement after ignoring those casts.

Note that we still have a problem if the base `@global` is also a
clause operand, for which the replacement happens before this constant
expr, because replacing `@global` inside he region would
also happen in the constant-expr `gep(@global, 0, 4)`. That is especially
an issue with opaque pointers. For that reason, the frontend will need to
emit Instructions instead of constant-exprs for clause operands
(CMPLRLLVM-42706), and `vpo-restore-operands` pass would need to be changed
to not restore constant-exprs into the region, but convert them into
instructions before replacement.

All this is because we are trying to make the IR treat clause operands as
distinct values, like function parameters, which cannot be used interchangeably,
but they are not function params.

JIRA: CMPLRLLVM-42010