[Driver] Remove support for conda triple (#10855) (#10866)
Support for the conda specific triple was added to allow for seamless
integration into the conda environment.  This change properly integrated
the toolchain into the environment.  Unfortunately, it is only partial
due to the fact that the conda libraries should be paired with the
sysroot that is supplied.  This sysroot is not easily accessible causing
a disconnect when compiling with the conda libraries due to the usage
of system based libraries as well (causing a glibc version mismatch).

Backout this change to allow for using the existing methodology of setting
both the --gcc-toolchain and --sysroot values when users want to use the
conda environment.  Default environment will continue to use whatever is
found on the system.

Addresses:  CMPLRLLVM-44800
Cherry pick of: https://github.com/intel-restricted/applications.compilers.llvm-project/commit/5295082e4af0fe52f277f6ddc4cffe310527440f