[HIRLoopInterchange] Suppress loop interchange for outer SIMD loops (#10948)
Found this issue while working on a different PR. Interchanging loops was
resulting in a runfail because we kept the outer loop as SIMD even after
interchange. We don't have the right semantic knowledge to allow interchange for
SIMD loops, assuming it is even allowed. This can probably be improved in the
future.