[OpenMP][Offload][L0] Memory pool fix and improvements (#10990)
JIRA: CMPLRLIBS-34330

Implemented the following changes in memory pool.

* Do not adjust max total pool size. This is due to an observation that
allowing too large allocation from pool results in huge slow down
in kernel execution. It seems that giving users explicit control of
this value is preferred in such cases.

* Create a space for dedicated small allocation after pool usage hits
the max total pool size. This is a request from CMPLRLIBS-34330, and
we allow small allocation size up to 1KB, and total size up to 4MB.

* Fix incorrect assertion in memory deallocation. Introduction of two
dedicated pools for reduction makes the existing assertion invalid, so
it was converted to a correct check.