[ScalarEvolution] Improve isImpliedCond() logic by building constant range for SCEV exprs (#10835)
This changeset introduces a ConstantRangeBuilder class which forms a
ConstantRange for LHS/RHS of a predicate to be able to reason about it by
starting with a known range formed using a dominating predicate.

This helps recognize more ZTTs and simplify loop trip count exprs. This
is needed for CMPLRLLVM-43678.

The relaxed loop logic in HIRRuntimeDD stopped working due to improved ZTT
recognition so it had to be revamped.