[VPlan] Track debug information for blends and reductions in HIR CG (#11043)
A customer complained about move instructions in a vectorized loop
being assigned the line number of the #pragma omp simd. These were
being generated as a result of blend operations. This patch fixes
this by identifying the non-header-phi input and copying the debug
location from there. It also updates the debug location correctly
for horizontal reductions in a similar manner. For now this only
fixes the problem for HIR code generation.

Reference: CMPLRLLVM-36270