[VPlan] Add basic support for UF selection in selectBestPlan (#10862)
Adds to selectBestPlan support to build a list of candidate UFs, and an inner loop in the main VF-selection loop over UF candidates. This means most of the changes in the main loop are whitespace or renaming 'BestUF' to the loop variable 'UF'.

Other changes of note:

    Checks for whether a forced/preferred VF and the previous 'BestUF' would exceed the trip count are replaced by code to prune all UFs for which the trip count would be exceeded, and check if the UFs list becomes empty. I have done my best to update the debug output to something sensible.
    The main loop had a map from VFs to the cost summary for that VF. The only use of that map was to look up the cost summary for the selected best VF. Rather than try to map VF/UF pairs , I have replaced it with a cost summary that is updated whenever we find a new best VF/UF. If there is any reason to keep such a map, please point it out.
    The corner case after the main selection loop where all costs are invalid was using the first available VF, and UF 1. I've changed this to use the first available VF and UF, where UF may be > 1 with -vplan-force-uf. If the intention here is to also ensure no unrolling is done, I can revert this line.

At present, the UFs list will only contain the result of getLoopUnrollFactor, i.e. a user-specified UF. There is also a facility for getLoopUnrollFactor to return a 'hint' from a previous pass, but I cannot find any case where this is used. For now I am distinguishing such a hint from a user-specified UF.