[VPlan] Fix interaction between unroller and VPlan transform that optimizes out single iteration vector loop (#11218)
When the combination of VF/UF results in a one-iteration vector loop, we try to
optimize the loop out at VPlan level in the LLVM IR path as later transforms
are not always able to clean up the same. Doing this too early was causing
couple of issues.

- trying to set branch latch condition during unroll in VPlan was hitting
an assert once the branch condition is changed to a constant(true/false).
- the code which tries to set the unroll factor in VectorTripCountCalculation
does not kick in once the latch condition is set to a constant.

To fix the issue, we deferred the change that optimizes out the loop to run after
unroller has run.

Part of changes to improve unroller stability.