[VPlan][DA]Change DA-shape calculation for a store. (#11206)
* [VPlan][DA]Change DA-shape calculation for a store.

In general, the DA-shape of a store does not matter as the instruction
does not have uses. Currently, we assign "joined" shapes of value and
pointer. That has absolutely no meaning, better to duplicate shape
of the pointer, at least we will not need to find it. The shape of the
value does not matter in the vast majority cases.