[VPlan][HIRCG] Fix assertion in HIR code generation of GEPs for pointer inductions with variable stride (#11263)
HIR code generation of GEPs for pointer inductions had a case where
it takes existing vector of pointers and indexes it with broadcast(step * VF).
While doing this, we need to create a copy instruction if the vector of
pointers is not a selfblob. Otherwise, we hit an assertion fail when
trying to get the selfblob index.

We hit this issue when we force unrolling the loop with pointer inductions.
Existing LIT test has been updated to check HIR for the unrolled case as
well.