[VPlan][CM] Add analysis and handling for no-cost instructions. (#10936)
Add a mechanism to the VPlan Cost Model to mark certain instructions as
having special behavior. This marking is done for instructions that
match specific patterns, during in an explicit analysis pass performed
during Cost Model construction.

The following special kinds are supported:
    - ZeroCost: always computed as zero-cost even if the instruction
      would otherwise be considered to have non-zero cost.

    - ZeroCostIfPeeling: same as ZeroCost, if cost modeling is queried
      within a peel loop.

The analysis currently recognizes a single pattern: assume
('@llvm.assume') operands. Instructions which are used only as an assume
operand, or by another instruction which is only used by an assume
operand, they are marked with 'ZeroCost'. This reflects the fact that
these instructions are not executed, and will be removed by later LLVM
passes.

JR: CMPLRLLVM-44397