[SYCL] Cherry-pick device_global patches (#11243)
* [SYCL] Implement device_global host-side memory operations (#8022)

This commit implements the copy and memcpy operations to and from
device_global. If the device_global does not have device_image_scope the
memory operation will be on the underlying USM memory, while if the
operation is on a device_global with device_image_scope the runtime will
try to find a suitable program in the program cache or build a new
program using the image using it.

* [SYCL][NFC] Ignore unused mock variable in device_global tests (#8491)

This commit prevents a warning due to unused mock variables in
device_global unittests. Note that the variable cannot be omitted as it
needs to stay alive for the duration of the test.

* [SYCL] Initialize members of device_global map entries (#8451)

This commit adds default initializers to select members of
DeviceGlobalMapEntry to avoid cases where bogus values would confuse
late initialization of the entries.

* [SYCL] Avoid passing pointer to temporary in device_global initialization (#8469)

This commit fixes a bug where the backend would be passed a pointer to a
temporary value holding the pointer to the address of the USM memory for
the device_global when initializing it, potentially causing wrong values
to be copied.

* [SYCL] Fix copy elision warning (#8437)

Fix post-commit:
https://github.com/intel/llvm/actions/runs/4244006981/jobs/7377519363

* [SYCL][NFC] Rename misleading queue_impl::MSupportOOO variable (#8342)

The variable is supposed to indicate that we failed to create an OOO
native queue and we need to emulate it if its value is false. Rename to
MEmulateOOO to avoid confusion with in-order/out-of-order SYCL queues.

---------

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>
Co-authored-by: Alexey Bader <alexey.bader@intel.com>
Co-authored-by: Dmitry Vodopyanov <dmitry.vodopyanov@intel.com>
Co-authored-by: Sergey Semenov <sergey.semenov@intel.com>