[cmp-43192] [RuntimeDD] Consider folded zero offset memrefs (#11184)
RuntimeDD doesn't try sorting memrefs in the same group when the shapes (#dims, offsets) are different.
For example, following group's elements have different shapes, and RuntimeDD avoids sorting. Delinearization is not successful either and runtimeDD test doesn't kick in.

Group 1 = { (%a)[0] and (%a)[0].1.0[1] }

However, all-zero indices or offsets are usually results of gep folding and the address of those memrefs are clear.
This PR make RuntimeDD sort a group even when element shapes' are not exactly the same when all-zero offset memrefs are element of the group. (All zero offset memrefs: a[0], a[0].0, and so on.)