[VPlan][HIR] Make the vector of private pointers DDRef consistent at current nesting level. (#11307)
We were currently setting the nesting level of the created
address-of ref to represent vector of private pointers as
candidate HLLoop's level - 1. However this would cause
invalid nesting level if this DDRef is used in another
HLNode prior to the inner vectorized loop - for example,
calls to constructors to initialize the pointer (non-POD
private). This PR fixes this discrepancy in def@level of
the DDRef by calling makeConsistent utility. We also
determine nesting level using insertion point instead of
hard-coded values.

JIRA: CMPLRLLVM-40468