[NFC] Use Type::getNonOpaquePointerElementType where applicable. (#11376)
Upstream LLVM deprecated Type::getPointerElementType and
PointerType::getElementType; due to the excessive number of warnings caused by
xmain's readiness for opaque pointers at the time, those deprecations were
removed in xmain customization.

This ports all of the queries of those two methods that are properly guarded by
a check for opaque pointers to use the non-deprecated, newer method. The
deprecation messages have not been restored, as there remains a body of OpenCL
code that has not yet been ported to support opaque pointers.