CMPLRLLVM-45658: Add explicit "make current" pass for classic inlining (#11469)
report.

A previous PR (#11319) causes compfails in post-commit testing of the
opt report. The problem is that right before the inlining report is
emitted, the IR may be in the process of being deleted. So, I created
a new pass to run 'make current' right before the metadata inlining
report does its emission.