Use O0 for debug build when building opencl libraries (#11502)
We always using O3 to build opencl runtime. This turns out to be very time consuming when we are doing a debug build,
when the opt is built with -O0, it is painful slow to use it to built runtime with O3.

Our pulldown bot is using debug build, and this is slowing down the build when doing incremental build.