[VPlan][PassManager] Run LoopSimplifyCFG after LowerSwitch, before VPlan. (#11378)
Add another instance of `LoopSimplifyCFGPass()` to the pipeline in
`PassManagerBuilder::addVPlanVectorizer()`, after LowerSwitch is run,
but before VPlanDriver. This is necessary because we run LowerSwitch
just before VPlan, but this pass can break loop-canonical form, which
is a precondition for VPlan. Consider the following failure case:

```llvm-ir
*** Before LowerSwitch:
<header>:
   %iv = phi i32 [ ..., <entry> ], [ ..., <latch> ]
   %exitcond = icmp ...
   br i1 %exitcond, <exit>, <latch>

<latch>:
   switch ... label <header> [
     ..., label <exit>
     ..., label <exit>
   ]

*** After LowerSwitch:
<header>:
   %iv = phi i32 [ ..., <entry> ], [ ..., <latch1> ], [ ..., <latch2> ]
   %exitcond = icmp ...
   br i1 %exitcond, label <exit>, label <latch1>

<latch1>:
   %case1.cond = icmp ...
   br i1 %case1.cond, label <header>, label <latch2>

<latch2>:
   %case2.cond = icmp ...
   br i1 %case2.cond, label <header>, label <exit>

*** Crash in VPlanDriver: multiple loop latches for loop
```

In order to prevent this, we can simply run LoopSimplifyCFG after
LowerSwitch to restore loop-canonical form for any loops which may
have had it broken by LowerSwitch.

JR: CMPLRLLVM-45477