[VPlan][NFCI] Add non-template base class to `VPPeelRemainderImpl`. (#11553)
Move the non-templated parts of `VPPeelRemainderImpl` (i.e. loop remarks
and needs-cloning flag) to a non-templated base class
`VPScalarLoopBase`. This allows us to reduce boilerplate in a few cases
where used to have to down-cast a `VPInstruction` to each derived class
just to call the base function, e.g:

    VPInstruction *I = /* ... */;
    if (auto *Lp = dyn_cast<VPScalarPeel>(I))
      Lp->setCloningRequired();
    else if (auto *Lp = dyn_cast<VPScalarPeelHIR>(I))
      Lp->setCloningRequired();
    else if (auto *Lp = dyn_cast<VPScalarRemainder>(I))
      Lp->setCloningRequired();
    else
      cast<VPScalarRemainderHIR>(I)->setCloningRequired();

Instead, we can just cast to `VPScalarLoopBase` and call the base class
function.