[VPlan][HIR] Handle SIMD descriptors with struct offsets in HIRLegality (#11305)
Patch to add checks in isSIMDDescriptorDDRef requiring that an incoming
reference match any struct field offsets present in the original SIMD descriptor,
if they exist. Previously, an incoming reference to the base struct pointer could
match to a reference to a field in the descriptor, leading to an InductionDescr
getting invalid information that would cause tryToCompleteByVPlan to fail.