[VPlan][HCFGBuilder][HCFGBuilderHIR] Fix for cmplrllvm-45685 (#11547)
* [VPlan][HCFGBuilder][HCFGBuilderHIR] Fix for cmplrllvm-45685

Step type for variable stride linear opaque pointers was set
incorrectly using the type size of the pointer. This lead to
an assertion when the stride multiplier is applied because
there were type mismatches for the operands of the mul. This
patch fixes the issue in the HCFGBuilders by using the type
from the step from the incoming LinearList item. This should
also prevent from having to insert some type conversion just
before the mul instruction.

* Clean up lit test