[OpenCL] Remove the __builtin_astype zeroinitializer for event_t. (#11602)
With the current OpenCL compiler, (event_t)(0) is a valid zeroinitializer for
the event type, which means that this method can be simplified. This further
avoids any issues with compiling the validation tests for SPIR-V.

This change is needed to enable support for target extension types for SPIR-V in
xmain, which is needed to enable opaque pointers. See also CMPLRLLVM-45796.