[VPlan][VLS][NFCI] Extend VLS transform related instructions to store group stride in granularity elements (#11705)
The changes lay the ground work for handling VLS groups such as a[2*i],
a[2*i+1], a[2*i+2] with no access gaps where the group stride(2) and group
size(3) do not match. Currently we only handle VLS groups where size and
absolute_value(stride) match. In order to support the cases when the size
and stride do not match, we need to use the stride and size information
in VLS transform and code generation. The changes in this PR start storing
stride information as well as part of VLS related VP instructions. Subsequent
changes will use the same to support VLS optimization for cases when stride
and size do not match.