[VPlan][NFC] Add unrolling heuristics for partial sums (not enabled) (#11399)
* [VPlan][NFC] Add unrolling heuristics for partial sums (not enabled).

The heuristic is to decrease the cost of the loop iteration to compensate the fact that compare instruction for main induction will be executed only once for all unrolled chunks. Also, the cost of reductions with multiple accumulators is accounted in a similar way - we consider that they all are executed in parallel.