[HIRLoopCollapse] Do not parse invalid AccumuBlobIndex in matchCEOnIVLevels(). (#11776)
* [HIRLoopCollapse] Do not parse invalid AccumuBlobIndex in matchCEOnIVLevels().

On each nesting level we try to match constant and blob indices of the IV with
accumulated constant and blob indices. For blob index, we also try to parse
accumulated TC blob to take into count possible types mismatch. The fix is to
bail out if accumulated TC blob do not match IV index and is invalid.