[DTRANS] Properly handle decomposed int128 types when passed directly (#11765)
As discovered in CMPLRLLVM-44967, when trying to arrange the type of a
function with a __int128_t parameter on Linux64 (where it is broken down
into 2 i64s), we hit an assert, as we didn't know how to form this.

This patch correctly handles that case. Additionally, it handles the
case of an int 128 in the return type, which on both platforms is some
sort of collection (either a literal struct, or a vector type), so
dtrans info doesn't matter.

note that 32 bit architectures aren't tested, because int 128 types
aren't enabled on those platforms.