[VPlan] Add opt-report bailout remarks for VPlan driver (#11828)
[VPlan] Add opt-report bailout remarks for VPlan driver

This is the first in a series of patches to improve optimization report
messages when the vectorizer is unable to vectorize a loop nest.  The
VPlan driver owns the mechanism for adding the remarks.  Future patches
for different vectorizer components will provide bail-out data from those
components for the driver to handle.