[Paropt] Tune new KNOWN_NDRANGE clause handling w/o specific partitioning enabled. (#11522)
The compiler attempts to help adjusting teams number a kernel is
launched with by passing the loop tripcounts to the RT even
when it's illegal to use specific NDRange. But that can hurt
performance for no the innermost loops, i.e. when the omp loop
is not the innermost one w/o collapse clause or when collapse
value is not enough to make it innermost.

JIRA: CMPLRLLVM-45632