Revert 11827 and cherry-pick three community commits. (#11853) (#11855)
* Revert "Resolve windows build check-all fail by revert two community commits (#11827)"

This reverts commit 4e769677d5842e469f72f8b0b31297468b0a3903.

* [support] Fix PrintNumber Test on AIX

When fixing the test earlier, we missed the JSON case for NaN and INF,
so handle those the same as for non-JSON, by creating the string
dynamically.

Reviewed By: abhina.sreeskantharajan

Differential Revision: https://reviews.llvm.org/D146739

* [support] Attempt to fix PrintNumber test for Solaris

NaN and Inf are still causing some problems in a formatting test.
This patch makes the checked format string exactly match the internal
JSON format string. If there are still problems, we should disable
testing Inf and NaN values until we can come to a portable solution.

Reviewed By: abhina.sreeskantharajan

Differential Revision: https://reviews.llvm.org/D146818

* [support] Temporarily disable Inf/NaN testing in PrintNumber

This is still breaking on some platforms. The underlying implementation
doesn't seem to be the cause, rather the test is not robust across
platforms. So, we'll just disable this for the time being, to unblock
builds until we have a proper fix.

Reviewed By: abhina.sreeskantharajan

Differential Revision: https://reviews.llvm.org/D146834
Fixes https://jira.devtools.intel.com/browse/CMPLRLLVM-46135
---------

Co-authored-by: Freddy Ye <freddy.ye@intel.com>
Co-authored-by: Paul Kirth <paulkirth@google.com>