[[ACD] Set resolver's "target-features" attribute to the feature-set supported by the "target-cpu" (#11814)
Compiling with flags like -mavx2, -mvpclmulqdq, -mavx512f, and -mvaes,
augments the "target-features" attribute of functions, with feature-names
that are not compatible with all target-cpu's the resolvers need to support.

Prior to this PR, AutoCPUClonePass set the "target-features" attribute of resolvers
by copying it over from the "target-features" attribute of the base/generic clones.
This resulted in vector instructions appearing in resolver functions. The vector
instructions failed on processors not supporting them. This PR fixes this by
explicitly setting the resolver's "target-features" attribute to the feature-set fully
supported by the "target-cpu". This PR fixes CMPLRLLVM-45676.