[CMPLRLLVM-43678] Add analysis for hir special opt predicate pass (#11635)
Adding pass to remove if that can be used calculated prior to loop execution.
The actual source code looks like the following:

// for (v = (-height/2 to height/2)
//   for (u = (-width/2 to width/2)
//     if ((v*v+u*u) <= (ssize_t) ((width/2)*(height/2)))
//       ...

Because v * v and u * u are positive and monotonicly increasing, for any value of
-/+ v, we can calculate the exact value between -u and -u where ((v * v+u * u) <=
size) holds. E.g. suppose size is 40 and v = 5, then the u loop must execute
only when u * u <= 40 - 5 * 5, which means u * u <= 15, so u spans [-3:+3]. Suppose
for the original loop, -u starts at any negative value. In this case, once
-u >= -3, we would enter the loop for that value of -u (from -u to u).
Otherwise we never enter the loop.

The transformation is the following:

for (v = (-height/2 to height/2))
  lb = 0
  ub = -1
  for (u = (-width/2 to width/2))
    if ((v*v+u*u) <= ((width/2)*(height/2)))
      lb = -u
      ub = u
      break;

  for (w = (lb to ub))
    ...