[VPlan][VLS] Apply VLS optimization for VLS groups with no gaps when stride and size are different (#11768)
* [VPlan][VLS] Apply VLS optimization for VLS groups with no gaps when stride and size are different

The changes in this PR apply VLS optimization for load  groups such as a[2i],
a[2i+1], a[2*i+2] with no access gaps. In this case, if we are vectorizing
for say VF == 4, we can do a wide load of 9 elements and generate appropriate
shuffles for the values corresponding to the individual loads. VLS analysis
is changed to pass through such groups. We still have the following
restrictions in place:

- for negative stride case, we need to shuffle values correctly. This will
follow in a later PR.
- stores typically have dependency issues. We need to see if we can do this
correctly.
- if the loads are under a mask, the mask for the wide load needs to be
computed appropriately. This is also left for later when we see a need.

I will add a LIT test to check for the above cases. I am posting this PR
to get the review process underway.

* The changes here start tracking if an OVLS memref is masked and avoid bailing
out of VLS altogether for groups such a[-2*i], a[-2*i+1], a[-2*i+2]. LIT
tests have been added showing VLS optimization kicking in as expected
for different cases.

* Assertion check was happening too early. Also avoided using the local MaskedAccessSeen flag.

* Address review comments

* clang-format change

* Address review comment