[VPlan] Add VPlan Legality bailout remarks to opt report (#11837)
This is the second patch in a series to add bailout reasons for
unvectorized loops to the optimization report.  This patch adds the
support within our loop vectorization legality framework and ties it
back to the VPlan driver support.  Some existing support for producing
bailout debug messages is removed, as it wasn't quite flexible enough
to work with the optimization report framework.