CMPLRLLVM-45774: Tolerate empty DTrans metadata for %rtti and %eh structs. (#11941)
* CMPLRLLVM-45774: Tolerate empty DTrans metadata for %rtti and %eh
structs.

The C/C++ front end does not want to provide DTrans metadata for RTTI and EH
structures. This caused a seg fault in the IRMover while compiling
523.xalancbmk with the Windows 32-bit compiler and VS=2022. Some modules tried
to use the DTrans type merging mechanism, others the standard type merging
mechanism. The solution is to tolerate a lack of DTrans metadata on structures
with "%rtti." and "%eh." name prefixes.

I needed to recreate this work area because running the FIF build on my
original workarea toasted it. The only difference between this and the
previous PR is that line 3160 of IRMover.cpp needs a comment:

or the FIF tool will not recognize the #else correctly.

* Clang format