[VPlan] Produce HIR bailout remark when LLVM-IR path will not be invoked (#11915)
In release compilers, we avoid placing vectorization bailout remarks in
the opt report from the HIR pass for SIMD loops.  This avoids customer
confusion with the remarks produced by the later LLVM-IR pass on the
same loop.  This patch allows producing the HIR bailout remarks for a
release compiler when we know the LLVM-IR pass has been disabled.