Cherry-pick "[SYCL] Fix accessor fill for unsupported patterns" (#12005)
* [SYCL] Fix accessor fill for unsupported patterns (#8845)

Currently most of the PI backends have similar pattern limitations as
OpenCL, i.e. only certain sizes of patterns are supported. This is not
currently being considered by the implementation however, potentially
causing the backend calls to fail. This commit makes the runtime fall
back to a fill kernel if the pattern has an unsupported size.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* [SYCL][NFC] Add missing Windows symbol (#8900)

This commit adds a missing Windows symbol to the symbols dump after
https://github.com/intel/llvm/pull/8845.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>