CMPLRLLVM-27323: 538.imagick PredicateOpt (#9): Hoistability of cache_info and nexus_info->region through the multiloop. (#12103)
* CMPLRLLVM-27323: 538.imagick PredicateOpt (#9): Hoistability of
cache_info and nexus_info->region through the multiloop.

In addition to proving that certain fields of cache_info and
nexus_info->region can be hoisted out of the wrapper function
@GetOneCacheViewVirtualPixel, we need to prove they can be hoisted out of
the (u,v) multiloop that surrounds the call to @GetOneCacheViewVirtualPixel
in @MeanShiftImage. There is one local struct variable that is written
in this multiloop. The comments show how we can prove that it is basically
used as a "scratchpad" and will not alias with the fields of cache_info
and nexus_info->region that we care about.

* Clang format

* Fix grammar mistake

* Fixes suggested by Prasad