[Paropt] Remove master-thread check from parallel global reduction. (#12082)
As the cross-team reduction stage is parallelized between the threads
of the single team it is executed by, there was an MTT for the write
to the resulting phi in each thread. That seems redundant as only
the master thread's result is written back to the reduction variable.
This change is needed for further refactoring of the atomic-free
reduction codegen.