CMPLRLLVM-29103 "Reenable DWARF Address Space And DW_OP_xderef Support" (#11989)
Enable the DWARF address space support in CLANG by removing the intel
customizations which disable it. The presence of this information breaks all
versions of GDB currently available. To work around this, a new debug pass
was created which removes no-op address spaces casts from the DWARF location
expressions. The effect is, we retain the address space information for
SPIRV/GPU targets while removing it from CPU targets.

JIRA: CMPLRLLVM-29103