538.imagick PredicateOpt (#10): Bug fix for previous change set (#12190)
* 538.imagick PredicateOpt (#10): Bug fix for previous change set

There are actually stores that are not to constants for the one struct
variable that is modified in the main loop. A few typos caused the
existence of the non-constant stores not to be noticed. It actually
does not matter that the values stored are not constant. We just
need to ensure that the address of the struct is not taken and stored
in a place which would cause it to invalidate the hoisting of
cache_info and nexus_info->region fields.

Please enter the commit message for your changes. Lines starting

* Fix spelling mistake caught by Praasad