Fix some coverity issues in clang codegen (#12180)
Coverity issue 2022794: In EmitBuiltinGenerateSIMDVariant getAs is used to
pull the FunctionProtoType out of the ElaboratedType. Use castAs so it
is clear the type must be found. This is ensured in Sema/checkVariantList.

Coverity issue 2770887: in EmitBuiltinCallSIMDVariant getAs is used to pull
the TemplateSpecializationType out of the ElaboratedType. Add as assert so
it is clear the type must be found. This is ensured in
CheckSIMDVariantBuiltinFunctionCall. Note that castAs doesn't work here
since that requires a CanonicalType and this is not.

Coverity issue 2769086: In OpenMPLateOutliner::OpenMPLateOutliner use cast
instead of dyn_cast. Since it is known that CurrentDirectiveKind is a loop
directive, the full directive 'D' must also be a loop directive.