[SYCL] Guard get_pointer API changes for 2023.2 (#12257)
The return type of the get_pointer member function was changed in
intel/llvm to adhere to the SYCL 2020 specification. However, this
API change was not intended for 2023.2, so this commit guards the
changes with the SYCL 2020 conformance macro (SYCL2020_CONFORMANT_APIS)
to avoid it inadverdently breaking user code for 2023.2. This commit
should be reverted for 2024.0.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>