[CMPLRLLVM-46532] Do not delete instructions with distpoint during Load Propagation (#12158)
Copy propagation was optimizing away a single use load that had distpoint attached to it. When Loop Distribution ran, it asserted that it could not find the directive as it was expecting one for that loop. The fix is in the Propagation util to ignore those nodes with the directive. It might be possible to delete the inst and move the directive, but that could lead to unintended cases so we just bail out instead.

JIRA: CMPLRLLVM-46532