review/skozhukh/soa select (#12120)
* [VPlan][SOA] Enable select instruction as a safe aliasing for privates

Curently, select is considered as preventing SOA layout for privates.
This change extends the safety analysis that we have for phis on selects
thus enabling SOA laout for private which pointers are used in select
instructions. Enabling is under the same conditions as for phis: select
with private and non-private or SOA-unsafe pointer-operands is prohibited.